requirements-app.txt
pyyaml
python-dateutil
readability-lxml
beautifulsoup4
lxml
feedparser
pandas
requests
